refactor(bootstrap): move non_gui_runtime into entrypoint/
Phase 4.3 of the uc-bootstrap redesign. Relocate non_gui_runtime.rs to
entrypoint/non_gui.rs, establishing the entrypoint/ subdir for scenario
constructors. Contract re-exports source from entrypoint::non_gui; the
LoggingHostEventEmitter reference in assembly.rs updated. No behavior change.
